+++ src/MathObjects/utils.cpp
@@ -1,6 +1,6 @@
 #include "utils.h"
 
-float distLineToPoint(Line line, Vector point)
+double distLineToPoint(Line line, Vector point)
 {
     return 1.1;
 }
